feat: Fix learn page, fix api, logout (#26)
온보딩 api, 학습 api 수정, 학습 페이지 수정(녹음 2번), 학습 ui 수정, toast에 string 적용, 로그아웃 api(미완), 로그인 유지 수정

+++ Android/app/src/main/java/com/example/pro_diction/data/dto/TestScoreDto.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class TestScoreDto(
     @SerialName("studyId") val studyId: Int,
-    @SerialName("score") val score: Float
+    @SerialName("score") val score: Int
 )
